requirements.txt
pandas
apache-airflow-providers-azure